Refact fetch functions to helper
