+++ backend/dump1090_collector/flight_storage.py
@@ -1,5 +1,3 @@
-# services/flight_storage.py
-
 from django.db.models import Q
 from django.utils import timezone
